[fixbug][ActionFilter] fix wrong when convert data from int? to int
Khi khai báo với kiểu generic K, nên sử khai báo type kiểu K("where K : struct") mới sử dụng được K? nếu không nó sẽ không hiểu được

+++ WebApiDapper/WebApiDapper/Controllers/ProductController.cs
@@ -33,7 +33,7 @@ namespace WebApiDapper.Controllers
 
         [HttpGet("id")]
         [ServiceFilter(typeof(ValidationNotExistEntityAttribute<Product, int>))]
-        public async Task<IActionResult> GetProductById(int id)
+        public IActionResult GetProductById(int id)
         {
             var product = HttpContext.Items["Entity"] as Product;
             if (product == null)
@@ -71,6 +71,7 @@ namespace WebApiDapper.Controllers
         public async Task<IActionResult> DeleteProduct(int id)
         {
             var product = HttpContext.Items["Entity"] as Product;
+            if (product == null) return NotFound();
             await _productService.DeleteProductByIdAsync(product.Id);
             return NoContent();
         }
